added validation to not allow delete category with devices

+++ src/services/category/category.service.ts
@@ -2,9 +2,14 @@ import { Injectable } from '@nestjs/common';
 import { Category } from 'src/models/category';
 import { CategoryRepository } from 'src/repositories/category/category.repository';
 
+import { DeviceService } from '../device/device.service';
+
 @Injectable()
 export class CategoryService {
-  constructor(private readonly categoryRepository: CategoryRepository) {}
+  constructor(
+    private readonly categoryRepository: CategoryRepository,
+    private readonly deviceService: DeviceService,
+  ) {}
 
   findAllCategories(): Promise<(Category | null)[]> {
     return this.categoryRepository.findAll();
@@ -18,7 +23,17 @@ export class CategoryService {
     return this.categoryRepository.createNew(category);
   }
 
-  deleteCategory(id: number): Promise<Category | null> {
+  async deleteCategory(id: number): Promise<Category | null> {
+    // prettier-ignore
+    const totalOfDevices = await this.deviceService.countDevicesByCategoryId(id);
+    const categoryHasDevices = totalOfDevices > 0;
+
+    if (categoryHasDevices) {
+      throw new Error(
+        'Could not delete this category because it still contains devices',
+      );
+    }
+
     return this.categoryRepository.delete(id);
   }
 }
